requirements.txt
aiodns==3.2.0
hikari==2.0.0.dev126
hikari_arc==1.3.4
httpx==0.27.2
python-dotenv==1.0.1
validators==0.33.0
